requirements-cuda.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121
transformers==4.48.0
accelerate==0.30.0
datasets==2.18.0
trl==0.17.0
peft>=0.12.0
numpy>=1.26
tqdm>=4.66
omegaconf>=2.3.0
ml-collections>=0.1.1
dm-tree>=0.1.8
biopython>=1.83
modelcif>=0.11
einops>=0.7.0
scipy>=1.10.0
packaging>=23.1
typing_extensions>=4.7.0
httpx>=0.24.0
wheel>=0.41.2
fair-esm @ git+https://github.com/facebookresearch/esm.git@main
git+https://github.com/aqlaboratory/openfold.git@v1.0.1

